pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iphone-screenshooter"
version = "0.1.0"
description = "A tool for exploring and capturing screenshots of iOS apps"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Jonathan Filippini", email = "jonathanfilippini@proton.me"},
]
dependencies = [
    "appium-python-client>=5.1.0",
]

[project.scripts]
ios-app-explorer = "ios_app_explorer.main:main"

[project.urls]
"Homepage" = "https://github.com/jonno85/iphone-screenshooter"
"Bug Tracker" = "https://github.com/jonno85/iphone-screenshooter/issues"

[tool.hatch.build.targets.wheel]
packages = ["ios_app_explorer"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
